🚚 Rename: 회원가입 페이지 파일 구조 변경 (#227)
